+++ src/roman_datamodels/testing/utils.py
@@ -369,6 +369,26 @@ def mk_dark(shape=None, filepath=None):
         return darkref
 
 
+def mk_gain(shape=None, filepath=None):
+    meta = {}
+    add_ref_common(meta)
+    gainref = stnode.GainRef()
+    meta['reftype'] = 'GAIN'
+    gainref['meta'] = meta
+
+    if shape:
+        gainref['data'] = np.zeros(shape, dtype=np.float32)
+    else:
+        gainref['data'] = np.zeros((4096, 4224), dtype=np.float32)
+
+    if filepath:
+        af = asdf.AsdfFile()
+        af.tree = {'roman': gainref}
+        af.write_to(filepath)
+    else:
+        return gainref
+
+
 def mk_mask(shape=True):
     meta = {}
     add_ref_common(meta)
@@ -382,6 +402,26 @@ def mk_mask(shape=True):
         maskref['dq'] = np.zeros((4096, 4096), dtype=np.uint32)
 
     return maskref
+
+
+def mk_readnoise(shape=None, filepath=None):
+    meta = {}
+    add_ref_common(meta)
+    readnoiseref = stnode.ReadnoiseRef()
+    meta['reftype'] = 'READNOISE'
+    readnoiseref['meta'] = meta
+
+    if shape:
+        readnoiseref['data'] = np.zeros(shape, dtype=np.float32)
+    else:
+        readnoiseref['data'] = np.zeros((4096, 4224), dtype=np.float32)
+
+    if filepath:
+        af = asdf.AsdfFile()
+        af.tree = {'roman': readnoiseref}
+        af.write_to(filepath)
+    else:
+        return readnoiseref
 
 
 def mk_ramp(arrays=True):