+++ tests/test_models.py
@@ -159,3 +159,39 @@ def test_opening_dark_ref(tmp_path):
     dark = datamodels.open(file_path)
     assert dark.meta.instrument.optical_element == 'F158'
     assert isinstance(dark, datamodels.DarkRefModel)
+
+# Gain tests
+def test_make_gain():
+    gain = utils.mk_gain(shape=(20,20))
+    assert gain.meta.reftype == 'GAIN'
+    assert gain.data.dtype == np.float32
+
+    # Test validation
+    gain_model = datamodels.RampModel(gain)
+    assert gain_model.validate() is None
+
+def test_opening_gain_ref(tmp_path):
+    # First make test reference file
+    file_path = tmp_path / 'testgain.asdf'
+    utils.mk_gain(filepath=file_path)
+    gain = datamodels.open(file_path)
+    assert gain.meta.instrument.optical_element == 'F158'
+    assert isinstance(gain, datamodels.GainRefModel)
+
+# Read Noise tests
+def test_make_readnoise():
+    readnoise = utils.mk_readnoise(shape=(20,20))
+    assert readnoise.meta.reftype == 'READNOISE'
+    assert readnoise.data.dtype == np.float32
+
+    # Test validation
+    readnoise_model = datamodels.RampModel(readnoise)
+    assert readnoise_model.validate() is None
+
+def test_opening_readnoise_ref(tmp_path):
+    # First make test reference file
+    file_path = tmp_path / 'testreadnoise.asdf'
+    utils.mk_readnoise(filepath=file_path)
+    readnoise = datamodels.open(file_path)
+    assert readnoise.meta.instrument.optical_element == 'F158'
+    assert isinstance(readnoise, datamodels.ReadnoiseRefModel)