+++ CHANGES.rst
@@ -1,7 +1,12 @@
 0.6.0 (unreleased)
 ==================
 
-- Added dark current model maker and tests. [#26]
+- Added Dark Current model maker and tests. [#26]
+
+- Added Readnoise maker utility and tests. [#23]
+
+- Added Gain maker utility and tests. [#24]
+
 
 0.5.0 (2021-08-07)
 ==================